pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "granita"
authors = [{ name = "apple-phi" }]
readme = "README.md"
dynamic = ["version", "description"]
requires-python = ">=3.6"
license = { file = "LICENSE" }
dependencies = [
    "markdown ~=3.3",
    "python-frontmatter ~=1.0",
    "pyquery ~=1.4",
    "typer ~=0.4"
]

[project.urls]
Home = "https://github.com/apple-phi/granita"

[project.scripts]
granita = "granita.__main__:app"
